+++ backend/database/taskStore.ts
@@ -15,7 +15,11 @@ export async function deleteTaskFromDB(task:TaskData) {
     const taskDataCollection = await returnCollection('taskData')
     const lastDocument = await returnLastDocument(taskDataCollection,{name:task.name})
     console.log("LAST DOCUMENT",lastDocument);
-    await taskDataCollection.deleteOne({ _id:lastDocument[0]._id })
+    if (lastDocument[0]) {
+        await taskDataCollection.deleteOne({ _id:lastDocument[0]._id })
+        return
+    }
+    await taskDataCollection.deleteOne({name:task.name})
 }
 export async function updateTaskToDB(originalTask:TaskData,newTask:TaskData) {
     const taskDataCollection = await returnCollection('taskData')
